Move debug.{h,cc} to src/firebuild/ and don't link debug.o with the interceptor
This fixes the build failure on Ubuntu 18.04 where the interceptor requires
PIC code.
